fix(misra): fix MISRA violations in Csm, Ethernet, MemIf, SoAd (final batch)
- Rule 17.7: cast unused return values to (void) in SoAd.c and Csm.c
- Rule 10.4: add U suffix to integer literals in SoAd.c, ethernet files
- Rule 21.15: add (void*)/(const void*) casts on memcpy calls in SoAd.c
- Rule 10.5: replace (boolean)(expr) with ternary ? TRUE : FALSE in SoAd.c
- Rule 2.7: suppress unused parameter warning in MemIf.c
- Rule 20.5: suppress #undef in ethernet_windows.c

Completes all 464 MISRA C:2012 violation fixes across the project.

+++ Autosar_SecOC/source/Csm/Csm.c
@@ -976,7 +976,7 @@ Std_ReturnType Csm_KeyElementSet(
         return CRYPTO_E_KEY_SIZE_MISMATCH;
     }
 
-    if ((uint32)CSM_KEYID_IS_SESSION(keyId) == (uint32)TRUE)
+    if (CSM_KEYID_IS_SESSION(keyId) == TRUE)
     {
         peerId = CSM_SESSION_PEER_FROM_KEYID(keyId);
         if (peerId >= CSM_MAX_PEERS)
@@ -1027,7 +1027,7 @@ Std_ReturnType Csm_KeySetValid(uint32 keyId)
         return E_OK;
     }
 
-    if ((uint32)CSM_KEYID_IS_SESSION(keyId) == (uint32)TRUE)
+    if (CSM_KEYID_IS_SESSION(keyId) == TRUE)
     {
         peerId = CSM_SESSION_PEER_FROM_KEYID(keyId);
         if (peerId >= CSM_MAX_PEERS)
